refactor(testing): move vtesting to internal/vtesting
This PR relocates the package github.com/kakao/varlog/vtesting to
github.com/kakao/varlog/internal/vtesting as it does not need to be exported.
The package supports flexibility in testing environments by providing features
like a test logger, assertion wait timeout, raft tick, and commit tick of the
metadata repository. However, these parameters are infrequently configured and
can be managed using Go's built-in testing package and helpers.

By moving it to an internal package, we aim to reduce the exposed API surface
and encourage using standard testing tools, eventually leading to the potential
deprecation of the github.com/kakao/varlog/internal/vtesting package. The
package is now being moved to the internal directory without any functional
changes.
