+++ pkg/varlog/trim_test.go
@@ -3,7 +3,7 @@ package varlog
 import (
 	"testing"
 
-	_ "github.com/kakao/varlog/vtesting"
+	_ "github.com/kakao/varlog/internal/vtesting"
 )
 
 func TestTrim(t *testing.T) {
